+++ src/main.h
@@ -20,22 +20,22 @@
 
  */
 
+extern Workspaceroot *main_workspaceroot;
+extern Toolkitgroup *main_toolkitgroup;
+extern Symbol *main_symbol_root;
+extern Watchgroup *main_watchgroup;
+extern Imageinfogroup *main_imageinfogroup;
+
+extern void *main_c_stack_base;
+
+extern gboolean main_option_time_save;
+extern gboolean main_option_profile;
+extern gboolean main_option_i18n;
+extern gboolean main_option_batch;
+extern gboolean main_option_no_load_menus;
+
+void set_prefix(const char *prefix);
 const char *get_prefix(void);
-void main_print_main(Symbol *sym);
 
-extern Workspaceroot *main_workspaceroot;	/* All the workspaces */
-extern Toolkitgroup *main_toolkitgroup;		/* All the toolkits */
-extern Symbol *main_symbol_root;			/* Root of symtable */
-extern Watchgroup *main_watchgroup;			/* All of the watches */
-extern Imageinfogroup *main_imageinfogroup; /* All of the images */
-
-extern void *main_c_stack_base;				/* Base of C stack */
-
-extern gboolean main_starting;				/* In startup */
-
-extern gboolean main_option_time_save;		/* Time save image ops */
-extern gboolean main_option_profile;		/* Profile calcualtion */
-extern gboolean main_option_i18n;			/* Output i18n strings */
-extern gboolean main_option_batch;			/* Running in batch mode */
-extern gboolean main_option_verbose;		/* Verbose output */
-extern gboolean main_option_print_main;		/* Print value of "main" on exit */
+void main_startup(int argc, char **argv);
+void main_shutdown(void);